I have removed the adapt_step and fixed_step separated algorithms that complicated so much the code (now I can integrate them properly, using the same structure but choosing what step & order predictor you want to use -> and the dense output would for example work for fixed step). I have also removed the constant checks on if it was dense. This is how it works (it is my main change besides the restructuring):
    The interpolating polynomial is only calculated if the user has
asked for a time in the (current_time,current_time+step) range. This
means that the extra function evaluations that were done at the time
current_time+step are not done unless interpolation is needed (this
should save some function evaluations at each step without interpolation
-exactly- (order+1) function evaluations)
    If the step taken doesn't need interpolation then the harmonic
sequence is used (so for the cases where only the final time is wanted
the algorithm is the same -> non-dense tests pass) if not the dense
sequence is used. This lays to different results for the regression
tests. I am currently forcing to take the dense sequence so it behaves
like the previous code and see if it passes the tests (it currently
doesn't). When it passes the dense tests too I am going to update them
so the tests work for the dense sequence adaptation.

The rejection of step works the same (but also saves the interpolation
evaluations at the final time if step is rejected)
There is a TODO on the symmetric vs non-symmetric methods use (change
the extrapolation formula basically).