Updated regression test to adapt this differences: Difference 1: The interpolating polynomial is only calculated if the user has asked for a time in the (current_time,current_time+step) range. This means that the extra function evaluations that were done at the time current_time+step are not done unless interpolation is needed (this should save some function evaluations at each step without interpolation -exactly- (order+1) function evaluations) If the step taken doesn't need interpolation then the harmonic sequence is used (so for the cases where only the final time is wanted the algorithm is the same -> non-dense tests pass) if not the dense sequence is used.
Difference 2:
    What code did: when the step you are taking falls exactly in a value
you wanted (for example at the final time this always happens), after
the step is not rejected (by the adaptation of order and step) you
interpolate the values (and maybe reject the step). The thing is that it
also interpolates this last value (although you have already the result
at that time, and accepted the error), and sometimes the error of
interpolation at this last value forces the rejection of the step.
    What this commit changes: you don't interpolate the value that you
already have and therefore the step is never rejected due to the
interpolation of the final value.
    
Also added some tests problems (only the problem, not actually testing
them)